- Moved sinusbot-section-manager into source dir - Changed sinusbot-scripting-engine to dev dependency - Created 'transpiler' which can import .js files inside a js function
