+++ package.json
@@ -1,7 +1,11 @@
 {
   "name": "sinusbot-section-manager",
   "version": "1.0.0",
-  "dependencies": {
+  "dependencies": {},
+  "devDependencies": {
     "sinusbot-scripting-engine": "^1.0.11"
+  },
+  "scripts": {
+    "build": "node transpileSinusbotScript.js"
   }
 }
